Move byline tests to htmlProcessing, stub formatter for tests
