+++ test/unit/htmlProcessing.test.js
@@ -2,12 +2,18 @@ const fs = require('fs')
 const path = require('path')
 const cheerio = require('cheerio')
 const { assert } = require('chai')
+const { getProcessedDocAttributes } = require('../../server/formatter')
+
+// helper function to stub the doc and get a section of the returned document
+function stubbedProcessedDoc(unprocessedHtml, editorName) {
+  const docData = {data: { lastModifyingUser: {displayName: editorName} }}
+  return getProcessedDocAttributes([unprocessedHtml, docData])
+}
 
 describe('HTML processing', () => {
   before(function () {
-    const formatter = require('../../server/formatter')
     this.rawHTML = fs.readFileSync(path.join(__dirname, '../fixtures/supportedFormats.html'), { encoding: 'utf8' })
-    this.processedHTML = formatter.getProcessedHtml(this.rawHTML)
+    this.processedHTML = stubbedProcessedDoc(this.rawHTML).html
     this.output = cheerio.load(this.processedHTML)
   })
 
@@ -95,6 +101,24 @@ describe('HTML processing', () => {
     it('strips inline comment anchors', function () {
       const commentAnchorParent = this.output("p:contains('will be stripped from the')")
       assert.notMatch(commentAnchorParent, /\[a\]/)
+    })
+  })
+
+  describe('byline fetching', () => {
+    it('should return reglar byline if none in HTML', () => {
+      const {byline} = stubbedProcessedDoc('<p></p>', 'Ben Koski')
+      assert.equal(byline, 'Ben Koski')
+    })
+
+    it('should return a byline if present in HTML', () => {
+      const {byline} = stubbedProcessedDoc('<p>By John Smith</p>', 'Ben Koski')
+      assert.equal(byline, 'John Smith')
+    })
+
+    it('should return byline override if present in document', () => {
+      const {byline} = stubbedProcessedDoc('<p>I am standing by Port Authority</p>', 'Ben Koski')
+      assert.notEqual(byline, 'Port Authority')
+      assert.equal(byline, 'Ben Koski')
     })
   })
 })
